AI_Dataset_Capture/requirements.txt
flask
picamera2
pyserial